Cards and Data Base
_Create Logic to horasMinistradas card
_Create Logic to Total de Alunos
_Create Logic to Aulas Disponiveis
_Rename Class DiaAulaDTO to  standardize
_Create Documentation to AlunoDAO Class
_Create Documentation to AulaDAO Class
_Create BuscarAluno in AlunoService Class
_Create Documentation to AlunoService Class
_Delete from the AdministradorDAO class due to code duplicity
_Add reference in button "Pagina Inicial"
_Ajust in redirect login
