+++ fluencyNow/src/com/br/fluencynow/controller/Cadastrar_controller.java
@@ -1,13 +1,10 @@
 package com.br.fluencynow.controller;
 
-import com.br.fluencynow.dao.AdministradorDAO;
 import com.br.fluencynow.dao.AlunoDAO;
 import com.br.fluencynow.dao.PlanoDAO;
 import com.br.fluencynow.dto.AlunoDTO;
 import com.br.fluencynow.model.Aluno;
 import com.br.fluencynow.model.Plano;
-import com.br.fluencynow.service.AlunoService;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -15,7 +12,6 @@ import org.springframework.web.servlet.ModelAndView;
 
 import javax.servlet.http.HttpSession;
 import java.sql.SQLException;
-import java.util.Date;
 import java.util.List;
 
 @Controller
@@ -24,12 +20,8 @@ public class Cadastrar_controller {
     @RequestMapping("/cadastrar")
     public ModelAndView cadastrar(Model model) {
 
-        List<Aluno> alunos = new AdministradorDAO().getAluno();
-        Aluno.Lista_container alunoList = new Aluno.Lista_container();
-        alunoList.setAlunos(alunos);
-        model.addAttribute("Alunos", alunoList);
-
         try {
+            new com.br.fluencynow.service.AlunoService().getAluno(model);
             new com.br.fluencynow.service.PlanoService().getPlano(model);
             new com.br.fluencynow.service.AulaService().getDia(model);
             new com.br.fluencynow.service.AulaService().getHora(model);
