+++ fluencyNow/src/com/br/fluencynow/controller/Login_controller.java
@@ -21,7 +21,7 @@ public class Login_controller {
     public String efetuaLogin(Login login, HttpSession session) throws SQLException {
         if(new com.br.fluencynow.dao.LoginDAO().exists(login)) {
             session.setAttribute("usuarioLogado", login);
-            return "administrador";
+            return "redirect:administrador";
         }
         return "redirect:login";
     }
