+++ fluencyNow/src/com/br/fluencynow/service/AlunoService.java
@@ -3,11 +3,17 @@ package com.br.fluencynow.service;
 import com.br.fluencynow.dao.AlunoDAO;
 import com.br.fluencynow.dto.AlunoDTO;
 import com.br.fluencynow.model.Aluno;
+import org.springframework.ui.Model;
 
 import java.sql.SQLException;
+import java.util.List;
 
 public class AlunoService {
 
+    /**
+     * Salva aluno no banco de dados
+     * @param aluno AlunoDTO
+     * */
     public boolean SalvarAluno(AlunoDTO aluno) throws SQLException {
 
         if(new com.br.fluencynow.validadores.ValidaCPF().validarCPF(aluno.cpf) == false){
@@ -21,6 +27,19 @@ public class AlunoService {
         return criarAluno;
     }
 
+    public String getAluno(Model model) throws SQLException{
+        List<Aluno> alunos = new AlunoDAO().getStudent();
+        Aluno.Lista_container alunoList = new Aluno.Lista_container();
+        alunoList.setAlunos(alunos);
+        model.addAttribute("Alunos", alunoList);
+
+        return "cadastrarAluno";
+    }
+    /**
+     * Deleta um aluno do banco de dados
+     * Deleta uma aula do banco de dados
+     * @param cpf String
+     * */
     public boolean DeletarAluno(String cpf) throws SQLException {
 
         Aluno aluno = new com.br.fluencynow.dao.AlunoDAO().searchIdStudentByCpf(cpf);
@@ -32,5 +51,45 @@ public class AlunoService {
         }
 
         return false;
+    }
+
+    /**
+     * Pega a quantidade de alunos no banco de dados e mostra na view
+     * @param model Model
+     * */
+    public String quantidadeAluno(Model model) throws SQLException {
+
+        int quantidade = AlunoDAO.getStudentCount();
+        model.addAttribute("quantidadeAluno", quantidade);
+        return "administrador";
+    }
+
+    /**
+     * Retorna a quantidade de aulas na view
+     * @param model Model
+     * */
+    public String aulasDisponiveis(Model model) throws SQLException {
+        int totalAulas = 45; //Multiplicação das horas disponiveis pelos dias de aula
+        int quantidade = AlunoDAO.getStudentCount();
+        int disponiveis = totalAulas - quantidade;
+
+        model.addAttribute("aulasDisponiveis", disponiveis);
+        return "administrador";
+    }
+
+    /**
+     * Retorna a quantidade de horasMinistradas na view
+     * @param model Model
+     * */
+    public String horasMinistradas(Model model) throws SQLException {
+        int horaAula = 0; //Hora Aula
+        int quantidade = AlunoDAO.getStudentCount();
+
+        for(int i = 1; i <= quantidade; i++){
+            horaAula += 45;
+        }
+
+        model.addAttribute("horasMinistradas", horaAula);
+        return "administrador";
     }
 }
