+++ fluencyNow/src/com/br/fluencynow/service/AulaService.java
@@ -2,7 +2,7 @@ package com.br.fluencynow.service;
 
 import com.br.fluencynow.dao.AulaDAO;
 import com.br.fluencynow.dto.AlunoDTO;
-import com.br.fluencynow.dto.DiaAula;
+import com.br.fluencynow.dto.DiaAulaDTO;
 import org.springframework.ui.Model;
 
 import java.util.ArrayList;
@@ -35,10 +35,10 @@ public class AulaService {
 
     public String getAulaEAluno(Model model){
 
-        List<DiaAula> listDiaAula = new ArrayList<>();
+        List<DiaAulaDTO> listDiaAulaDTO = new ArrayList<>();
 
         /*Lista das 08H*/
-        listDiaAula.add(new DiaAula("08:00",
+        listDiaAulaDTO.add(new DiaAulaDTO("08:00",
                 new AulaDAO().getClassAndStudent(DiasdaSemana.SEGUNDA_FEIRA.toString(), HorarioAula.Aula_08h00.toString()),
                 new AulaDAO().getClassAndStudent(DiasdaSemana.TERCA_FEIRA.toString(), HorarioAula.Aula_08h00.toString()),
                 new AulaDAO().getClassAndStudent(DiasdaSemana.QUARTA_FEIRA.toString(), HorarioAula.Aula_08h00.toString()),
@@ -46,7 +46,7 @@ public class AulaService {
                 new AulaDAO().getClassAndStudent(DiasdaSemana.SEXTA_FEIRA.toString(), HorarioAula.Aula_08h00.toString())));
 
         /*Lista das 09H*/
-        listDiaAula.add(new DiaAula("09:00",
+        listDiaAulaDTO.add(new DiaAulaDTO("09:00",
                 new AulaDAO().getClassAndStudent(DiasdaSemana.SEGUNDA_FEIRA.toString(), HorarioAula.Aula_09h00.toString()),
                 new AulaDAO().getClassAndStudent(DiasdaSemana.TERCA_FEIRA.toString(), HorarioAula.Aula_09h00.toString()),
                 new AulaDAO().getClassAndStudent(DiasdaSemana.QUARTA_FEIRA.toString(), HorarioAula.Aula_09h00.toString()),
@@ -54,7 +54,7 @@ public class AulaService {
                 new AulaDAO().getClassAndStudent(DiasdaSemana.SEXTA_FEIRA.toString(), HorarioAula.Aula_09h00.toString())));
 
         /*Lista das 10H*/
-        listDiaAula.add(new DiaAula("10:00",
+        listDiaAulaDTO.add(new DiaAulaDTO("10:00",
                 new AulaDAO().getClassAndStudent(DiasdaSemana.SEGUNDA_FEIRA.toString(), HorarioAula.Aula_10h00.toString()),
                 new AulaDAO().getClassAndStudent(DiasdaSemana.TERCA_FEIRA.toString(), HorarioAula.Aula_10h00.toString()),
                 new AulaDAO().getClassAndStudent(DiasdaSemana.QUARTA_FEIRA.toString(), HorarioAula.Aula_10h00.toString()),
@@ -62,7 +62,7 @@ public class AulaService {
                 new AulaDAO().getClassAndStudent(DiasdaSemana.SEXTA_FEIRA.toString(), HorarioAula.Aula_10h00.toString())));
 
         /*Lista das 11H*/
-        listDiaAula.add(new DiaAula("11:00",
+        listDiaAulaDTO.add(new DiaAulaDTO("11:00",
                 new AulaDAO().getClassAndStudent(DiasdaSemana.SEGUNDA_FEIRA.toString(), HorarioAula.Aula_11h00.toString()),
                 new AulaDAO().getClassAndStudent(DiasdaSemana.TERCA_FEIRA.toString(), HorarioAula.Aula_11h00.toString()),
                 new AulaDAO().getClassAndStudent(DiasdaSemana.QUARTA_FEIRA.toString(), HorarioAula.Aula_11h00.toString()),
@@ -70,7 +70,7 @@ public class AulaService {
                 new AulaDAO().getClassAndStudent(DiasdaSemana.SEXTA_FEIRA.toString(), HorarioAula.Aula_11h00.toString())));
 
         /*Lista das 12H*/
-        listDiaAula.add(new DiaAula("12:00",
+        listDiaAulaDTO.add(new DiaAulaDTO("12:00",
                 new AulaDAO().getClassAndStudent(DiasdaSemana.SEGUNDA_FEIRA.toString(), HorarioAula.Aula_12h00.toString()),
                 new AulaDAO().getClassAndStudent(DiasdaSemana.TERCA_FEIRA.toString(), HorarioAula.Aula_12h00.toString()),
                 new AulaDAO().getClassAndStudent(DiasdaSemana.QUARTA_FEIRA.toString(), HorarioAula.Aula_12h00.toString()),
@@ -78,7 +78,7 @@ public class AulaService {
                 new AulaDAO().getClassAndStudent(DiasdaSemana.SEXTA_FEIRA.toString(), HorarioAula.Aula_12h00.toString())));
 
         /*Lista das 13H*/
-        listDiaAula.add(new DiaAula("13:00",
+        listDiaAulaDTO.add(new DiaAulaDTO("13:00",
                 new AulaDAO().getClassAndStudent(DiasdaSemana.SEGUNDA_FEIRA.toString(), HorarioAula.Aula_13h00.toString()),
                 new AulaDAO().getClassAndStudent(DiasdaSemana.TERCA_FEIRA.toString(), HorarioAula.Aula_13h00.toString()),
                 new AulaDAO().getClassAndStudent(DiasdaSemana.QUARTA_FEIRA.toString(), HorarioAula.Aula_13h00.toString()),
@@ -86,7 +86,7 @@ public class AulaService {
                 new AulaDAO().getClassAndStudent(DiasdaSemana.SEXTA_FEIRA.toString(), HorarioAula.Aula_13h00.toString())));
 
         /*Lista das 15H*/
-        listDiaAula.add(new DiaAula("15:00",
+        listDiaAulaDTO.add(new DiaAulaDTO("15:00",
                 new AulaDAO().getClassAndStudent(DiasdaSemana.SEGUNDA_FEIRA.toString(), HorarioAula.Aula_15h00.toString()),
                 new AulaDAO().getClassAndStudent(DiasdaSemana.TERCA_FEIRA.toString(), HorarioAula.Aula_15h00.toString()),
                 new AulaDAO().getClassAndStudent(DiasdaSemana.QUARTA_FEIRA.toString(), HorarioAula.Aula_15h00.toString()),
@@ -94,7 +94,7 @@ public class AulaService {
                 new AulaDAO().getClassAndStudent(DiasdaSemana.SEXTA_FEIRA.toString(), HorarioAula.Aula_15h00.toString())));
 
         /*Lista das 16H*/
-        listDiaAula.add(new DiaAula("16:00",
+        listDiaAulaDTO.add(new DiaAulaDTO("16:00",
                 new AulaDAO().getClassAndStudent(DiasdaSemana.SEGUNDA_FEIRA.toString(), HorarioAula.Aula_16h00.toString()),
                 new AulaDAO().getClassAndStudent(DiasdaSemana.TERCA_FEIRA.toString(), HorarioAula.Aula_16h00.toString()),
                 new AulaDAO().getClassAndStudent(DiasdaSemana.QUARTA_FEIRA.toString(), HorarioAula.Aula_16h00.toString()),
@@ -102,15 +102,15 @@ public class AulaService {
                 new AulaDAO().getClassAndStudent(DiasdaSemana.SEXTA_FEIRA.toString(), HorarioAula.Aula_16h00.toString())));
 
         /*Lista das 17H*/
-        listDiaAula.add(new DiaAula("17:00",
+        listDiaAulaDTO.add(new DiaAulaDTO("17:00",
                 new AulaDAO().getClassAndStudent(DiasdaSemana.SEGUNDA_FEIRA.toString(), HorarioAula.Aula_17h00.toString()),
                 new AulaDAO().getClassAndStudent(DiasdaSemana.TERCA_FEIRA.toString(), HorarioAula.Aula_17h00.toString()),
                 new AulaDAO().getClassAndStudent(DiasdaSemana.QUARTA_FEIRA.toString(), HorarioAula.Aula_17h00.toString()),
                 new AulaDAO().getClassAndStudent(DiasdaSemana.QUINTA_FEIRA.toString(), HorarioAula.Aula_17h00.toString()),
                 new AulaDAO().getClassAndStudent(DiasdaSemana.SEXTA_FEIRA.toString(), HorarioAula.Aula_17h00.toString())));
 
-        DiaAula.Lista_container diaAula = new DiaAula.Lista_container();
-        diaAula.setDiaAula(listDiaAula);
+        DiaAulaDTO.Lista_container diaAula = new DiaAulaDTO.Lista_container();
+        diaAula.setDiaAula(listDiaAulaDTO);
         model.addAttribute("DiaAula", diaAula);
 
         return "administrador";
